cerberus/requirements.txt
Django==3.2
djangorestframework==3.13.0
psycopg2==2.9.5
requests==2.28.1
django-allauth==0.51.0
python-jose==3.3.0
python-decouple==3.7
git+https://github.com/Peter-Slump/django-keycloak.git
redis==4.5.1